+++ src/test/java/org/idnt/udemy/springboot/app/example/test/controller/AccountControllerWebTestClientTest.java
@@ -21,6 +21,7 @@ import java.util.Map;
 import static org.hamcrest.Matchers.*;
 import static org.junit.jupiter.api.Assertions.*;
 
+@Tag("integration-webTestClient")
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class AccountControllerWebTestClientTest {